+++ src/components/Background/index.js
@@ -1,0 +1,7 @@
+import BackgroundBlob from "./BackgroundBlob";
+import BackgroundBlobRepeat from "./BackgroundBlobRepeat";
+
+export {
+    BackgroundBlob,
+    BackgroundBlobRepeat
+}
