+++ src/components/Card/index.js
@@ -1,0 +1,11 @@
+import FeedCard from "./FeedCard";
+import MyOfferCard from "./MyOfferCard";
+import MyRequestCard from "./MyRequestCard";
+import OfferCard from "./OfferCard";
+
+export {
+    FeedCard,
+    MyOfferCard,
+    MyRequestCard,
+    OfferCard
+}
