+++ src/components/FormInput/index.js
@@ -1,0 +1,7 @@
+import FormInput from "./FormInput";
+import ImageInput from "./ImageInput";
+
+export{
+    FormInput,
+    ImageInput
+}
