+++ src/pages/Forms/NewOffer.jsx
@@ -1,12 +1,14 @@
 import React from 'react'
 import Style from './NewOffer.module.scss'
-import PageHeader from '../../components/PageHeader/PageHeader'
-import BackgroundBlob from '../../components/Background/BackgroundBlobRepeat'
-import {MultiStepForm, Step} from '../../components/MultiStepForm/MultiStepForm'
+import {
+    PageHeader,
+    BackgroundBlob,
+    MultiStepForm, Step,
+    FormInput, ImageInput,
+    MyRequestCard as Card
+} from '../../components'
 import {useRequestFormStore} from '../../store/FormStore'
 import {useForm} from 'react-hook-form'
-import {FormInput, ImageInput} from '../../components/FormInput/FormInput'
-import Card from '../../components/Card/MyRequestCard'
